@@ -1,60 +1,56 @@
 {
   "name": "judgepad",
   "version": "1.0.0",
+  "type": "module",
   "description": "Simple starter package for Redux with React and Babel support",
   "main": "index.js",
   "repository": "git@github.com:rlaakkol/judgepad.git",
   "scripts": {
-    "start": "NODE_ENV=development node ./node_modules/webpack-dev-server/bin/webpack-dev-server.js",
-    "test": "mocha --compilers js:babel-core/register --require ./test/test_helper.js --recursive ./test",
-    "test:watch": "npm run test -- --watch",
-    "test:lint": "eslint ./src --ext .js --ext .jsx --cache",
-    "package": "NODE_ENV=production webpack --progress --colors",
-    "deploy": "npm run package && surge -p $(git rev-parse --show-toplevel)"
+    "start": "vite",
+    "test": "vitest --config ./vite.config.mjs",
+    "test:watch": "vitest --config ./vite.config.mjs",
+    "test:lint": "ESLINT_USE_FLAT_CONFIG=false eslint ./src --ext .js --ext .jsx",
+    "lint:inspect": "eslint-config-inspector",
+    "build": "vite build",
+    "serve": "vite build && vite preview"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel-core": "^6.2.1",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.1.18",
-    "babel-preset-react": "^6.1.18",
-    "chai": "^3.5.0",
-    "chai-jquery": "^2.0.0",
-    "eslint": "^3.12.2",
-    "eslint-config-airbnb": "^13.0.0",
-    "eslint-config-google": "^0.7.1",
-    "eslint-config-prettier": "^1.6.0",
-    "eslint-loader": "^1.6.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-prettier": "^2.0.1",
-    "eslint-plugin-react": "^6.8.0",
-    "jquery": "^2.2.1",
-    "jsdom": "^8.1.0",
-    "mocha": "^2.4.5",
-    "prettier": "^1.1.0",
-    "react-addons-test-utils": "^15.4.2",
-    "webpack": "^1.12.9",
-    "webpack-dev-server": "^1.14.0"
+    "@babel/eslint-parser": "^7.27.5",
+    "@babel/preset-env": "^7.27.2",
+    "@babel/preset-react": "^7.27.1",
+    "@eslint/config-inspector": "^1.1.0",
+    "@eslint/js": "^9.30.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@vitejs/plugin-react": "^4.6.0",
+    "@babel/core": "^7.27.7",
+    "eslint": "^9.30.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "globals": "^16.2.0",
+    "jquery": "^3.7.1",
+    "prettier": "^3.6.2",
+    "redux-mock-store": "^1.5.5",
+    "vite": "^6.0.0",
+    "vite-plugin-pwa": "1.0.0",
+    "vitest": "^3.2.4"
   },
   "dependencies": {
-    "babel-preset-stage-1": "^6.1.18",
-    "lodash": "^3.10.1",
-    "marked": "^0.3.6",
-    "node-uuid": "^1.4.7",
-    "prop-types": "^15.5.8",
-    "react": "^15.4.2",
-    "react-bootstrap": "^0.30.7",
-    "react-dom": "^15.4.2",
-    "react-marked": "^0.3.1",
-    "react-redux": "^4.0.0",
-    "react-router": "^2.0.1",
-    "react-router-bootstrap": "^0.23.1",
-    "redux": "^3.0.4",
-    "redux-storage": "^4.1.2",
-    "redux-storage-engine-localstorage": "^1.1.4",
-    "surge": "^0.20.1"
+    "bootstrap": "^5.3.7",
+    "dompurify": "^3.2.6",
+    "lodash": "^4.17.21",
+    "marked": "^16.0.0",
+    "prop-types": "^15.8.1",
+    "react": "^19.1.0",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^19.1.0",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^7.6.3",
+    "redux": "^5.0.1",
+    "redux-persist": "^6.0.0",
+    "uuid": "^11.1.0"
   }
 }
